moved most of main into a method so I could call it again if they want to play again
